simplify process_notifications example; move example with notifications callback to the folder for advanced usage
